caption more slides + fix stage empty text + common stage output textslides
